feat: universal MinUI distribution with per-device input mappings
Ship both grout32 and grout64 in a single MinUI zip. The launch
script detects arch via uname and picks the right binary and libs.
Device detection (Miyoo/Anbernic/TrimUI) uses runtime.GOARCH and
/proc/bus/input/devices, following the same pattern as muOS. Adds
Anbernic joystick mapping reused from muOS.
