+++ cfw/minui/input_mappings/anbernic.json
@@ -1,0 +1,25 @@
+{
+  "keyboard_map": {},
+  "controller_button_map": {},
+  "controller_hat_map": {},
+  "joystick_axis_map": {},
+  "joystick_button_map": {
+    "10": 13,
+    "11": 15,
+    "12": 10,
+    "13": 12,
+    "3": 5,
+    "4": 6,
+    "5": 8,
+    "6": 7,
+    "7": 9,
+    "8": 11,
+    "9": 14
+  },
+  "joystick_hat_map": {
+    "1": 1,
+    "2": 4,
+    "4": 2,
+    "8": 3
+  }
+}
