+++ taskfiles/package.yml
@@ -4,7 +4,7 @@ tasks:
 
   all:
     desc: Package for all platforms
-    deps: [ next, muos, knulli, spruce, rocknix, trimui, allium, onion, minui, minui-arm32, batocera, batocera-x86, batocera-amd64 ]
+    deps: [ next, muos, knulli, spruce, rocknix, trimui, allium, onion, minui, batocera, batocera-x86, batocera-amd64 ]
     cmds:
       - echo "Packaging complete (13 platforms)"
     silent: true
@@ -94,20 +94,15 @@ tasks:
 
   minui:
     cmds:
-      - rm -rf dist/MinUI-arm64
-      - mkdir -p dist/MinUI-arm64/Grout/lib
-      - cp build64/grout scripts/MinUI/launch.sh README.md LICENSE dist/MinUI-arm64/Grout
-      - cp -R build64/lib/* dist/MinUI-arm64/Grout/lib/
-      - chmod a+x dist/MinUI-arm64/Grout/grout dist/MinUI-arm64/Grout/launch.sh
-    silent: true
-
-  minui-arm32:
-    cmds:
-      - rm -rf dist/MinUI-arm32
-      - mkdir -p dist/MinUI-arm32/Grout/lib
-      - cp build32/grout scripts/MinUI/launch.sh README.md LICENSE dist/MinUI-arm32/Grout
-      - cp -R vendored/miyoo/* dist/MinUI-arm32/Grout/lib/
-      - chmod a+x dist/MinUI-arm32/Grout/grout dist/MinUI-arm32/Grout/launch.sh
+      - rm -rf dist/MinUI
+      - mkdir -p dist/MinUI/Grout/lib64
+      - mkdir -p dist/MinUI/Grout/lib32
+      - cp build64/grout dist/MinUI/Grout/grout64
+      - cp build32/grout dist/MinUI/Grout/grout32
+      - cp scripts/MinUI/launch.sh README.md LICENSE dist/MinUI/Grout
+      - cp -R build64/lib/* dist/MinUI/Grout/lib64/
+      - cp -R vendored/miyoo/* dist/MinUI/Grout/lib32/
+      - chmod a+x dist/MinUI/Grout/grout64 dist/MinUI/Grout/grout32 dist/MinUI/Grout/launch.sh
     silent: true
 
   batocera:
